Revert "feat(web): add invitecode to homepage and login"
This reverts commit 69daa001ed669d85b2d7744d085370232d0727b5.
